jshint: add dependency
jslint complains about the polyfill for Object.assign and provides no
mechanism for disabling itself around a block of code.  I'm not going to
"fix" the code, so we're going to use a more configurable tool.  Leave
jslint as an alternative script until I'm sure jshint finds the problems
I care about.