Union: support variable lengths using absent default layout
We don't support default layouts with variable length because the length
would be part of the content and would be unlikely to be compatible with
any other variants.  So use a null default layout to indicate that the
union as a whole has variable length, determined by the variant it
holds.

Kinda complex, probably should be explained better.