Layout: support determining span of layout by data length
A primary use case of this module is encoding and decoding packetized
data, where the length of the packet can implicitly determine the length
of a member element.  Add GreedyCount as an external layout producing
the number of elements for which space is available at an offset within
the buffer, so we can leverage the implicit length.