Layout: remove Object.freeze() on layout instances
The original motivation for this was to ensure that coupled property
state (such as span and fields) could not be inadvertently made
inconsistent.  But freezing layouts this way complicates the binding of
layouts to prototypes for objects into which the layout is stored.  As a
step to simplifying that binding remove the freeze operation.