Layout: prefer interpretation as property over msb in BitStructure
BitStructure's constructor takes two parameters, with the second being a
rarely-used option to order allocation from the most significant bit
down.  The developer forgets this, and passes a property name without a
second argument, which results in no data.  If the argument is
type-compatible with property and no third argument is present, treat
the argument as property rather than msb.

Technically this is an API change, but I'm going to rule it minor
because anybody who's affected will have been using it wrong.

Closes #17.