travis.yml: remove older node versions from test
mocha has a dependency on yargs which only supports actively
maintained Node.js releases, which means they outlaw 6 and 8.  At the
moment they're OK with 10, but only because they haven't released
since it reached EOL on 2021-04-30.

Drop the versions that won't build in travis as of today.  We'll have
to test manually to confirm old Node.js release pass.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>